requirements.txt
matplotlib==3.5.3
numpy==1.26.4
pandas==1.5.0
Pillow==9.2.0
Pillow==10.2.0
plotly==5.18.0
scikit_learn==1.4.0
seaborn==0.13.2
streamlit==1.30.0
